Move `LoadConfig` from `config/__init__.py` to `config/load.py`
