+++ tests/model_executor/model_loader/test_registry.py
@@ -4,7 +4,8 @@
 import pytest
 from torch import nn
 
-from vllm.config import LoadConfig, ModelConfig
+from vllm.config import ModelConfig
+from vllm.config.load import LoadConfig
 from vllm.model_executor.model_loader import (get_model_loader,
                                               register_model_loader)
 from vllm.model_executor.model_loader.base_loader import BaseModelLoader
